+++ src/redux/shop/shop.actions.js
@@ -23,15 +23,15 @@ export const fetchCollectionsFailure = (errorMessage) => ({
 
 export const fetchCollectionsStartAsync = () => {
   return (dispatch) => {
-    const collectionRef = firestore.collection("collections");
-    dispatch(fetchCollectionsStart());
+  //   const collectionRef = firestore.collection("collections");
+  //   dispatch(fetchCollectionsStart());
 
-    collectionRef
-      .get()
-      .then((snapshot) => {
-        const collectionsMap = convertCollectionsSnapshotToMap(snapshot);
-        dispatch(fetchCollectionsSuccess(collectionsMap));
-      })
-      .catch((error) => dispatch(fetchCollectionsFailure(error.message)));
-  };
+  //   collectionRef
+  //     .get()
+  //     .then((snapshot) => {
+  //       const collectionsMap = convertCollectionsSnapshotToMap(snapshot);
+  //       dispatch(fetchCollectionsSuccess(collectionsMap));
+  //     })
+  //     .catch((error) => dispatch(fetchCollectionsFailure(error.message)));
+  // };
 };
