+++ src/redux/shop/shop.sagas.js
@@ -1,9 +1,30 @@
 import { takeEvery } from "redux-saga/effects";
 
+import {
+  firestore,
+  convertCollectionsSnapshotToMap,
+} from "../../Firebase/firebase.utils";
+
+import {
+  fetchCollectionsSuccess,
+  fetchCollectionsFailure,
+} from "./shop.actions";
+
 import ShopActionTypes from "./shop.types";
 
 export function* fetchCollectionsAsync() {
   yield console.log("I am fired");
+
+  const collectionRef = firestore.collection("collections");
+  dispatch(fetchCollectionsStart());
+
+  collectionRef
+    .get()
+    .then((snapshot) => {
+      const collectionsMap = convertCollectionsSnapshotToMap(snapshot);
+      dispatch(fetchCollectionsSuccess(collectionsMap));
+    })
+    .catch((error) => dispatch(fetchCollectionsFailure(error.message)));
 }
 
 export function* fetchCollectionsStart() {
